requirements.txt
streamlit
pandas
sqlite-utils
psycopg2-binary
nicegui

